Forward scroll events up the responder chain.
The default implementation of a responder is to forward scroll events up
the chain.

If a TUIScrollView cannot scroll in the given direction, forward it up
the chain.

Accomplish the direction detection by inspecting the direction of the
first NSEventPhaseChanged.

Include an example of nested scroll views within the example project.